Rename new path, replace existing one
